chore: consolidate config files
