--- src/config/config.dev.ts
@@ -1,228 +1,0 @@
-import rWETHIcon from '@/assets/token-logos/rWETH.png'
-import kWETHIcon from '@/assets/token-logos/kWETH.png'
-import gWETHIcon from '@/assets/token-logos/gWETH.png'
-import USDTIcon from '@/assets/token-logos/USDT.png'
-import USDCIcon from '@/assets/token-logos/USDC.png'
-import DAIIcon from '@/assets/token-logos/DAI.png'
-// import XDAIIcon from '@/assets/token-logos/XDAI.png'
-// import DEVIcon from '@/assets/token-logos/DEV.png'
-// import wADAIcon from '@/assets/token-logos/wADA.png'
-// import wEvmosIcon from '@/assets/token-logos/wEVMOS.png'
-import ConnextIcon from '@/assets/icons/connext.svg'
-
-
-import testnetTokens from './tokens.dev'
-import { TokenMetadata } from './config.types'
-
-export const tokens: { [key: string]: TokenMetadata } = {
-  // Only for use with connext in dev environment
-  kTEST: {
-    nativeNetwork: 'kovan',
-    symbol: 'kTEST',
-    name: 'Kovan TEST',
-    icon: ConnextIcon,
-    iconColors: ['#62BBEF', '#8470E2'],
-    decimals: 18,
-    coinGeckoId: 'dai',
-    tokenIdentifier: testnetTokens.TEST,
-    nativeOnly: false,
-    minAmt: 10,
-  },
-  WETH: {
-    nativeNetwork: 'rinkeby',
-    symbol: 'WETH',
-    name: 'Rinkeby WETH',
-    icon: rWETHIcon,
-    iconColors: ['#eda532', '#eda532'],
-    decimals: 18,
-    coinGeckoId: 'weth',
-    tokenIdentifier: testnetTokens.WETH,
-    nativeOnly: false,
-    minAmt: 0.0028,
-  },
-  ETH: {
-    nativeNetwork: 'rinkeby',
-    symbol: 'ETH',
-    name: 'Rinkeby ETH',
-    icon: rWETHIcon,
-    iconColors: ['#eda532', '#eda532'],
-    decimals: 18,
-    coinGeckoId: 'ethereum',
-    tokenIdentifier: null,
-    nativeOnly: true,
-    minAmt: 0.0028,
-    wrappedAsset: 'WETH',
-  },
-  kWETH: {
-    nativeNetwork: 'kovan',
-    symbol: 'kWETH',
-    name: 'Kovan WETH',
-    icon: kWETHIcon,
-    iconColors: ['#9B59B6', '#9B59B6'],
-    decimals: 18,
-    coinGeckoId: 'weth',
-    tokenIdentifier: testnetTokens.kWETH,
-    nativeOnly: false,
-    minAmt: 0.0028,
-  },
-  kETH: {
-    nativeNetwork: 'kovan',
-    symbol: 'kETH',
-    name: 'Kovan ETH',
-    icon: kWETHIcon,
-    iconColors: ['#9B59B6', '#9B59B6'],
-    decimals: 18,
-    coinGeckoId: 'ethereum',
-    tokenIdentifier: null,
-    nativeOnly: true,
-    minAmt: 0.0028,
-    wrappedAsset: 'kWETH',
-  },
-  gWETH: {
-    nativeNetwork: 'goerli',
-    symbol: 'gWETH',
-    name: 'Goerli WETH',
-    icon: gWETHIcon,
-    iconColors: ['#2980B9', '#2980B9'],
-    decimals: 18,
-    coinGeckoId: 'weth',
-    tokenIdentifier: testnetTokens.gWETH,
-    nativeOnly: false,
-    minAmt: 0.0028,
-  },
-  gETH: {
-    nativeNetwork: 'goerli',
-    symbol: 'gETH',
-    name: 'Goerli ETH',
-    icon: gWETHIcon,
-    iconColors: ['#2980B9', '#2980B9'],
-    decimals: 18,
-    coinGeckoId: 'ethereum',
-    tokenIdentifier: null,
-    nativeOnly: true,
-    minAmt: 0.0028,
-    wrappedAsset: 'gWETH',
-  },
-  USDT: {
-    nativeNetwork: 'kovan',
-    symbol: 'USDT',
-    name: 'USDT',
-    icon: USDTIcon,
-    iconColors: ['#fff', '#26A17B'],
-    decimals: 6,
-    coinGeckoId: 'tether',
-    tokenIdentifier: testnetTokens.USDT,
-    nativeOnly: false,
-    minAmt: 10,
-  },
-  USDC: {
-    nativeNetwork: 'kovan',
-    symbol: 'USDC',
-    name: 'USDC',
-    icon: USDCIcon,
-    iconColors: ['#fff', '#2976C9'],
-    decimals: 6,
-    coinGeckoId: 'usd-coin',
-    tokenIdentifier: testnetTokens.USDC,
-    nativeOnly: false,
-    minAmt: 10,
-  },
-  DAI: {
-    nativeNetwork: 'kovan',
-    symbol: 'DAI',
-    name: 'DAI',
-    icon: DAIIcon,
-    iconColors: ['#f0bf54', '#F4B731'],
-    decimals: 18,
-    coinGeckoId: 'dai',
-    tokenIdentifier: testnetTokens.DAI,
-    nativeOnly: false,
-    minAmt: 10,
-  },
-  // XDAI: {
-  //   nativeNetwork: 'xdai',
-  //   symbol: 'XDAI',
-  //   name: 'XDAI',
-  //   icon: XDAIIcon,
-  //   iconColors: ['#03795B', '#03795B'],
-  //   decimals: 18,
-  //   coinGeckoId: 'xdai',
-  //   tokenIdentifier: null,
-  //   nativeOnly: true,
-  //   minAmt: 10,
-  // },
-  // WXDAI: {
-  //   nativeNetwork: 'xdai',
-  //   symbol: 'WXDAI',
-  //   name: 'WXDAI',
-  //   icon: XDAIIcon,
-  //   iconColors: ['#03795B', '#03795B'],
-  //   decimals: 18,
-  //   coinGeckoId: 'wrapped-xdai',
-  //   tokenIdentifier: testnetTokens.WXDAI,
-  //   nativeOnly: false,
-  //   minAmt: 10,
-  // },
-  // DEV: {
-  //   nativeNetwork: 'moonbasealpha',
-  //   symbol: 'DEV',
-  //   name: 'Moonbase DEV',
-  //   icon: DEVIcon,
-  //   iconColors: ['#53CBC8', '#e84195'],
-  //   decimals: 18,
-  //   coinGeckoId: 'moonbeam',
-  //   tokenIdentifier: null,
-  //   nativeOnly: true,
-  //   minAmt: 10,
-  // },
-  // milkADA: {
-  //   nativeNetwork: 'milkomedaC1testnet',
-  //   symbol: 'mADA',
-  //   name: 'milkADA',
-  //   icon: wADAIcon,
-  //   iconColors: ['#6684CD', '#0033AC'],
-  //   decimals: 18,
-  //   coinGeckoId: 'cardano',
-  //   tokenIdentifier: null,
-  //   nativeOnly: true,
-  //   minAmt: 10,
-  //   wrappedAsset: 'wADA',
-  // },
-  // wADA: {
-  //   nativeNetwork: 'milkomedaC1testnet',
-  //   symbol: 'wADA',
-  //   name: 'wADA',
-  //   icon: wADAIcon,
-  //   iconColors: ['#6684CD', '#0033AC'],
-  //   decimals: 18,
-  //   coinGeckoId: 'cardano',
-  //   tokenIdentifier: testnetTokens.wADA,
-  //   nativeOnly: false,
-  //   minAmt: 10,
-  // },
-  // tEVMOS: {
-  //   nativeNetwork: 'evmostestnet',
-  //   symbol: 'tEVMOS',
-  //   name: 'tEVMOS',
-  //   icon: wEvmosIcon,
-  //   iconColors: ['#454545', '#000'],
-  //   decimals: 18,
-  //   coinGeckoId: 'evmos',
-  //   tokenIdentifier: null,
-  //   nativeOnly: true,
-  //   minAmt: 10,
-  //   wrappedAsset: 'wtEVMOS',
-  // },
-  // wtEVMOS: {
-  //   nativeNetwork: 'evmostestnet',
-  //   symbol: 'wtEVMOS',
-  //   name: 'wtEVMOS',
-  //   icon: wEvmosIcon,
-  //   iconColors: ['#454545', '#000'],
-  //   decimals: 18,
-  //   coinGeckoId: 'evmos',
-  //   tokenIdentifier: testnetTokens.wtEVMOS,
-  //   nativeOnly: false,
-  //   minAmt: 10,
-  // },
-}
