+++ src/store/modules/wallet.ts
@@ -9,7 +9,7 @@ import { networks } from '@/config/index'
 import * as mmUtils from '@/utils/metamask'
 import { getNetworkByChainID, nullToken } from '@/utils'
 import { TokenIdentifier } from '@nomad-xyz/sdk-bridge'
-import { MainnetNetwork, TestnetNetwork } from '@/config/config.types'
+import { MainnetNetwork, TestnetNetwork } from '@/config/types'
 
 export interface WalletState {
   connected: boolean
