+++ src/views/Transfer/Input/Input.amount.vue
@@ -94,7 +94,7 @@ import { helpers } from '@vuelidate/validators'
 
 import { useStore } from '@/store'
 import { getMinAmount, toDecimals } from '@/utils'
-import { TokenMetadata } from '@/config/config.types'
+import { TokenMetadata } from '@/config/types'
 import TokenSelect from './Input.tokens.vue'
 
 interface ComponentData {
